requirements.txt
deepface==0.0.95
flask==3.1.2
flask-cors==6.0.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
gunicorn==23.0.0
h5py==3.14.0
joblib==1.5.2
keras==3.11.3
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
requests==2.32.5
spotipy==2.25.1
streamlit==1.50.0
streamlit-webrtc==0.63.5
tenacity==9.1.2
tensorflow==2.20.0
python-dotenv==1.0.0  
